azure_functions/blob_upload_trigger/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.13.2
azure_storage_blob==12.14.1
rasterio==1.3.5.post1
rio_cogeo==3.5.0